Bitti, kontroller yapılıyor.

+++ Project.COREMVC/Models/RecipeDetails/RequestModels/RecipeDetailRequestModel.cs
@@ -3,9 +3,9 @@
     public class RecipeDetailRequestModel
     {
         public int RecipeID { get; set; }
-        public int IngredientID { get; set; }
+        public int IngredientID{ get; set; }
         public string Instruction { get; set; }
-        public int IngredientQuantity { get; set; }
+        public decimal IngredientQuantity { get; set; }
         public string Unit { get; set; }
 
     }
