+++ Project.ENTITIES/Models/Ingredient.cs
@@ -10,9 +10,9 @@ namespace Project.ENTITIES.Models
     public class Ingredient: BaseEntity
     {
         public string Name { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public decimal UnitPrice { get; set; }
-        public int PredictedAmount { get; set; }
+        public decimal PredictedAmount { get; set; }
         public string Unit { get; set; }
 
         //Relational Properties
