+++ Project.ENTITIES/Models/RecipeDetails.cs
@@ -11,7 +11,7 @@ namespace Project.ENTITIES.Models
         public int RecipeID { get; set; }
         public int IngredientID { get; set; }
         public string Instruction { get; set; }
-        public int IngredientQuantity { get; set; }
+        public decimal IngredientQuantity { get; set; }
         public string Unit { get; set; }
       
         //Relational Properties
